TP-1027: Convert to multi-module project
* This is done in a separate commit in order to keep git history for moved files
* `gs-test` artifact is relocated to `gs-test-junit4`
* Introduced new artifact `gs-test-core`, containing most of the earlier available code
* `gs-test-junit4` is intended to contain the junit4-specific code
* Introduced `gs-test-parent` to make this a multi-module project
