+++ gs-test-core/src/main/java/com/avanza/gs/test/JVMGlobalGigaSpacesManager.java
