+++ gs-test-core/src/main/resources/pu.xml.template
