+++ gs-test-core/src/main/resources/zoo.cfg.template
