+++ gs-test-core/src/test/java/com/avanza/gs/test/PuXmlEmulationTest.java
