+++ gs-test-core/src/test/java/com/avanza/gs/test/helpers/FruitPojo.java
