+++ gs-test-core/src/test/resources/fruit-mirror-pu.xml
