+++ gs-test-core/src/test/resources/fruit-pu.xml
