+++ gs-test-core/src/test/resources/log4j2.xml
