+++ gs-test/pom.xml
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+		 xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+		 xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+	<modelVersion>4.0.0</modelVersion>
+	<parent>
+		<groupId>com.avanza.gs</groupId>
+		<artifactId>gs-test-parent</artifactId>
+		<version>2.1.0-SNAPSHOT</version>
+	</parent>
+
+	<artifactId>gs-test</artifactId>
+
+	<distributionManagement>
+		<relocation>
+			<artifactId>gs-test-junit4</artifactId>
+			<message>The content of this artifact has been moved to gs-test-junit4</message>
+		</relocation>
+	</distributionManagement>
+
+</project>
